pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "chromadb>=1.3.7",
    "fastapi>=0.128.0",
    "ipykernel>=7.1.0",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.5",
    "langchain-google-genai>=4.1.2",
    "pymupdf>=1.26.7",
    "pypdf>=6.5.0",
    "python-multipart>=0.0.21",
    "sentence-transformers>=5.2.0",
    "vertexai>=1.71.1",
]
